requirements.txt
pyrogram
tgcrypto
aiohttp
psutil
yt-dlp
python-dotenv